+++ ruby/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add support for bundle Apache HTTP Components Proxy Configuration shinesolutions/aem-aws-stack-builder#235
 
 ### Changed